requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-Mail==0.9.1
itsdangerous==2.1.2
google-generativeai
folium==0.14.0
geopy==2.3.0
pdfkit==1.0.0
psycopg[binary]==3.2.9
Werkzeug==2.3.6
python-dotenv==1.0.0
rapidfuzz==3.13.0
waitress
